+++ src/test/java/com/solvd/twitter/service/UserServiceTest.java
@@ -1,12 +1,11 @@
-package com.solvd.twitter.unit;
+package com.solvd.twitter.service;
 
 import com.solvd.twitter.domain.exception.ResourceAlreadyExistsException;
 import com.solvd.twitter.domain.exception.ResourceDoesNotExistException;
 import com.solvd.twitter.domain.user.User;
 import com.solvd.twitter.repository.UserRepository;
-import com.solvd.twitter.service.UserService;
 import com.solvd.twitter.service.impl.UserServiceImpl;
-import com.solvd.twitter.unit.factory.TwitterFactory;
+import com.solvd.twitter.service.factory.TwitterFactory;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
